Add - 채팅 기능 추가 (#140)
* feat: websocket dependency 추가

* feat: 배틀 참가자가 있으면 삭제할 수 없도록 변경

* feat: chatting schema 추가

* feat: websocket connect authentication

* feat: websocket config

* todo: websocket event listener

* feat: chatting domain

* feat: 배틀 참여 및 참여취소 시 이벤트에 memberId 추가

* feat: 채팅 전송 및 배틀 참여/취소 시 상태채팅 전송 기능 추가

* feat: message type column 이름 type 으로 변경

* todo: test chatting service

* feat: 채팅 생성날짜 추가, 조회 시 생성날짜 역순으로 정렬

* feat: chatting topic battles 로 변경

+++ src/main/java/com/poorlex/poorlex/battle/battle/service/BattleService.java
@@ -322,9 +322,17 @@ public class BattleService {
     @Transactional
     public void delete(Long memberId, Long battleId) {
         validateAuthority(battleId, memberId);
+        validateMemberCanDeleteBattle(battleId);
 
         battleRepository.deleteById(battleId);
         Events.raise(new BattleDeletedEvent(battleId));
+    }
+
+    private void validateMemberCanDeleteBattle(Long battleId) {
+        if (battleParticipantRepository.findAllByBattleId(battleId)
+                .size() > 1) {
+            throw new BadRequestException(ExceptionTag.BATTLE_PARTICIPANT_SIZE, "참가자가 있는 배틀은 삭제할 수 없습니다.");
+        }
     }
 
     @Transactional
